Add support for `Prio3Aes128CountVec`
Allows testing of the `Prio3Aes128CountVec` VDAF. We add it to the
common structure for `VDAF`, and amend the definitions of the upload
request and collect poll response messages so that the measurement and
aggregate result (respectively) are defined generically for VDAFs, and
unconditionally encode numbers as strings. The goal here is to allow
representing `Prio3Aes128CountVec` measurements, but also to allow this
protocol to work with new VDAFs with minimal or no changes.

We still explicitly discuss specific VDAFs to define encoding task
parameters.